+++ bofire/data_models/strategies/api.py
@@ -30,11 +30,11 @@ from bofire.data_models.strategies.stepwise.stepwise import (  # noqa: F401
     Step,
     StepwiseStrategy,
 )
-from bofire.data_models.strategies.stepwise.transforms import (  # noqa: F401
+from bofire.data_models.strategies.strategy import Strategy
+from bofire.data_models.transforms import (  # noqa: F401
     AnyTransform,
     RemoveTransform,
 )
-from bofire.data_models.strategies.strategy import Strategy
 
 AbstractStrategy = Union[
     Strategy,
